Refactors shell script used to start and ssh into instance; adds script to transfer jupyter notebook from EC2 to local machine.
The script that used to start instance and ssh into it was broken up into 2 separate files.
This fixes the problem that previously the script had to be run twice, because the first time only started
the instance but resulted in a "port 22 closed" error, even after allowing additional time after starting the instance.